Disallow final actions in looping regex
Currently, Automa allows you to add a :final action to a regex like "a*", where
it is not possible to determine when reading a byte whether it is the final
byte matching the regex.
Current behaviour is to execute the action for every byte that could concievably
be the final byte whether or not it is.

Disallow this behaviour by throwing an error in re2nfa for :final actions in
looping regex.